writing DataType values to Tensor removed, DataStore updated, tests updated.
