+++ mithril/framework/physical/data_store.py
@@ -12,13 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from collections.abc import Mapping, Sequence
+from collections.abc import Callable, Mapping, Sequence
 from copy import deepcopy
 from typing import Any, TypeGuard
 
 from ...backends.backend import Backend
 from ...core import DataType, data_types
-from ...utils.func_utils import is_make_array_required, prepare_function_args
+from ...utils.func_utils import is_make_array_required
 from ...utils.utils import BiMap
 from ..common import (
     TBD,
@@ -33,6 +33,8 @@ from ..common import (
     is_type_adjustment_required,
 )
 from .flat_graph import FlatGraph
+
+key_map_type = dict[str, str]
 
 
 class StaticDataStore(GenericDataType[DataType]):
@@ -49,6 +51,9 @@ class StaticDataStore(GenericDataType[DataType]):
 
         self.is_materialized = False
         self._all_data: dict[str, Tensor[DataType] | Scalar] = dict()
+        # TODO: Combine data_values and tensor_values into one attribute and
+        # remove _materialize_cached_data method.
+        self._tensor_values: dict[str, DataType] = dict()
         self.data_memo: dict[int, Tensor[DataType] | Scalar] = dict()
         self.graph: FlatGraph[DataType] = graph
         self.backend: Backend[DataType] = backend
@@ -136,7 +141,7 @@ class StaticDataStore(GenericDataType[DataType]):
         )
         for data in updated_inter_data:
             key = self._intermediate_non_differentiables.inverse[data]
-            if data.value is not TBD:
+            if self.get_value(key) is not TBD:
                 if key in self._cached_data:
                     raise KeyError(
                         f"'{key}' key can not be an intermediate and cached key "
@@ -175,9 +180,10 @@ class StaticDataStore(GenericDataType[DataType]):
     def _materialize_cached_data(self):
         # Simply assigns final (real) values to the keys.
         if not self.data_values:
-            for key, data in self._cached_data.items():
-                assert not isinstance(data.value, ToBeDetermined)
-                self.data_values[key] = data.value
+            for key in self._cached_data:
+                value = self.get_value(key)
+                assert not isinstance(value, ToBeDetermined)
+                self.data_values[key] = value
             self.is_materialized = True
         else:
             raise Exception("Can not materialize cached data multiple times.")
@@ -280,6 +286,9 @@ class StaticDataStore(GenericDataType[DataType]):
             # Any fixes?.
             if isinstance(data, Tensor) and self.is_tensor_type(value):
                 updates |= data.set_value(value)
+                # Temporarily remove value from tensor and add to tensor_values!
+                self._tensor_values[key] = value
+                data.value = TBD
             elif isinstance(data, Scalar) and self.is_scalar_type(value):
                 updates |= data.set_value(value)  #
             else:
@@ -344,20 +353,19 @@ class StaticDataStore(GenericDataType[DataType]):
                     key: value
                     for key, value in zip(local_input_keys, value_mapping, strict=False)
                 }
-                args_dict, kwargs_dict = prepare_function_args(
-                    self.all_data,
+                args_dict, kwargs_dict = self.prepare_function_args(
                     fn,
                     inputs,
                     self.backend.array_creation_funcs,
                     False,
                 )
                 args = [
-                    statics[arg_key].value
+                    self.get_value(arg_key)
                     for arg_keys in args_dict.values()
                     for arg_key in arg_keys
                 ]
                 kwargs = {
-                    key: statics[value].value for key, value in kwargs_dict.items()
+                    key: self.get_value(value) for key, value in kwargs_dict.items()
                 }
 
                 # If function needs backend specific args
@@ -381,10 +389,139 @@ class StaticDataStore(GenericDataType[DataType]):
                     if is_make_array_required(data):
                         static_value = self.backend.array(static_value)
 
-                # queue |= self.add_static_data(value, static_value)
                 _queue, _updates = self.add_static_data(value, static_value)
                 queue |= _queue
                 updates |= _updates
         # Finalize cached_data.
         self._materialize_cached_data()
         return updates
+
+    def get_value(self, key: str) -> DataType | MainValueType | ToBeDetermined | str:
+        return self._tensor_values.get(key, self._all_data[key].value)
+
+    def prepare_function_args(
+        self,
+        function: Callable[..., Any],
+        inputs: key_map_type,
+        array_creation_funcs: list[str],
+        reduce_with_defaults: bool = True,
+    ) -> tuple[dict[str, list[str]], key_map_type]:
+        formula_key = function.__name__
+        code_obj = function.__code__
+
+        fn_args_dict = {
+            arg_name: False for arg_name in code_obj.co_varnames[: code_obj.co_argcount]
+        }
+
+        # Check if variadic positional argument exists
+        if code_obj.co_flags & 0x04 == 0x04:
+            idx = code_obj.co_argcount + code_obj.co_kwonlyargcount
+            fn_args_dict[code_obj.co_varnames[idx]] = True
+
+        fn_kwarg_keys = list(
+            code_obj.co_varnames[
+                code_obj.co_argcount : code_obj.co_argcount + code_obj.co_kwonlyargcount
+            ]
+        )
+
+        # Array creation functions requires device and
+        # precision to properly create tensor
+        if formula_key in array_creation_funcs:
+            # Remove precision and device from kwarg_keys
+            # we will partially provide them
+            fn_args_dict.pop("precision", None)
+            if "precision" in fn_kwarg_keys:
+                fn_kwarg_keys.remove("precision")
+
+            fn_args_dict.pop("device", None)
+            if "device" in fn_kwarg_keys:
+                fn_kwarg_keys.remove("device")
+
+        # Prepare arguments
+        fn_kwarg_dict, removed_kwarg_dict = self.create_kwarg_dict(
+            fn_kwarg_keys, function, inputs, reduce_with_defaults
+        )
+        fn_args_mapping = self.reorganize_args(
+            fn_args_dict,
+            set(fn_kwarg_dict.values()) | set(removed_kwarg_dict.values()),
+            function,
+            inputs,
+            reduce_with_defaults,
+        )
+
+        return fn_args_mapping, fn_kwarg_dict
+
+    def create_kwarg_dict(
+        self,
+        kwarg_keys: list[str],
+        function: Callable,
+        inputs: key_map_type,
+        reduce_with_defaults: bool,
+    ) -> tuple[key_map_type, key_map_type]:
+        kwarg_keys_dict: key_map_type = {
+            kwarg_key: inputs[kwarg_key] for kwarg_key in kwarg_keys
+        }
+        removed_kwargs_dict: key_map_type = {}
+
+        kwdefaults = function.__kwdefaults__
+
+        if kwdefaults is not None and reduce_with_defaults:
+            for key, value in kwdefaults.items():
+                # provided_value = data[kwarg_keys_dict[key]].value
+                provided_value = self.get_value(kwarg_keys_dict[key])
+                if value == provided_value and type(value) is type(provided_value):
+                    removed_kwargs_dict[key] = kwarg_keys_dict[key]
+                    kwarg_keys_dict.pop(key)
+
+        return kwarg_keys_dict, removed_kwargs_dict
+
+    def reorganize_args(
+        self,
+        arg_keys: dict[str, bool],
+        kwarg_keys: list[str] | set[str],
+        function: Callable,
+        inputs: key_map_type,
+        reduce_with_defaults: bool,
+    ) -> dict[str, list[str]]:
+        defaults = function.__defaults__
+        formula_key = function.__name__
+
+        local_input_keys = list(inputs.keys())
+        inputs = deepcopy(inputs)
+        organized_arguments: dict[str, list[str]] = {}
+
+        for idx, (name, is_variadic) in enumerate(arg_keys.items()):
+            if "cache" in name:
+                # TODO: Refactor here
+                provided_value = self.get_value(inputs[name])
+                if (
+                    reduce_with_defaults
+                    and idx == len(arg_keys) - 1
+                    and defaults
+                    and provided_value == defaults[-1]
+                ):
+                    continue
+
+                outer_names = [inputs[name]]
+                inputs.pop(name)
+
+            # If the argument variadic, then it takes rest of the inputs
+            elif is_variadic:
+                outer_names = [
+                    input for input in inputs.values() if input not in kwarg_keys
+                ]
+
+            elif name not in local_input_keys:
+                raise RuntimeError(
+                    f"Primitive '{formula_key}' input keys:'{local_input_keys}' and"
+                    f" backend function input keys: '{arg_keys.keys()}' "
+                    "are not matching!"
+                )
+
+            else:
+                outer_names = [inputs[name]]
+                inputs.pop(name)
+
+            organized_arguments[name] = outer_names
+
+        return organized_arguments
